Update 09 - Legion.lua
I followed the format of the shadowlands achievement, I added in all of the CRS #s and coords for the legion version of the achievement. Feel free to decline if I did something wrong with it, I have been talking to Runawaynow on discord about this. If you do decline this can you let me know why and what I did wrong so I can fix it, If everything is right, I can have the remainder of these Achievements fixed by this time next week